+++ reeln_google_plugin/__init__.py
@@ -2,7 +2,7 @@
 
 from __future__ import annotations
 
-__version__ = "0.7.0"
+__version__ = "0.8.0"
 
 from reeln_google_plugin.plugin import GooglePlugin